Update bedrock-entry.sh to clean up old settings
Setting `white-list` to true with no `whitelist.json` file causes all players to show up as not invited. If this container was ran with a version prior to https://github.com/itzg/docker-minecraft-bedrock-server/pull/337/files using an allow/white list, `white-list=true` would be set in `server.properties` in the Docker volume, but PR 337 removed the `whitelist.json` file.