Update to allow | white List handling
This proposal does the following:

If neither ALLOW_LIST_USERS or WHITE_LIST_USERS env variables are set, the allowlist.json is removed and ALLOW_LIST is exported as a false value.

If both of  variables are set, ALLOW_LIST_USERS takes precedence and WHITE_LIST_USERS is ignored.

If only one of these variables are set, it will function as expected.

ALLOW_LIST is the only exported variable, which should be sufficient for server.properties. It's also exported outside of the if block.